requirements.txt
asgiref==3.11.1
dj-database-url==3.1.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
psycopg2==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.5
